% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_imd_score.R
\name{calculate_imd_score}
\alias{calculate_imd_score}
\title{Calculate IMD score from domain scores}
\usage{
calculate_imd_score(
  scores_tbl,
  domains = choose_domains(),
  weights = imd_weights()
)
}
\arguments{
\item{scores_tbl}{A table containing transformed domain scores, as produced
by \link{get_transformed_scores}}

\item{domains}{A logical vector, best produced by \link{choose_domains}}

\item{weights}{A numeric vector, as produced by \link{imd_weights} by default}
}
\description{
You may exclude 1 or more domains by switching the relevant element of the
output of \link{choose_domains} to \code{FALSE}. You may additionally supply your own
vector of weights. See \link{imd_weights} or the IMD documentation for the
default weight values.
Setting a domain weight to zero is equivalent to excluding the domain, and
vice versa.
}
