% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_imd_score.R
\name{choose_domains}
\alias{choose_domains}
\title{Helper function to select domains to include in calculation of IMD overall}
\usage{
choose_domains(
  include_income = TRUE,
  include_employment = TRUE,
  include_education = TRUE,
  include_health = TRUE,
  include_crime = TRUE,
  include_barriers = TRUE,
  include_environment = TRUE
)
}
\arguments{
\item{include_income}{Boolean, default \code{TRUE}. Include the Income Deprivation
domain?}

\item{include_employment}{Boolean, default \code{TRUE}. Include the Employment
Deprivation domain?}

\item{include_education}{Boolean, default \code{TRUE}. Include the Education,
Skills and Training domain?}

\item{include_health}{Boolean, default \code{TRUE}. Include the Health
Deprivation and Disability domain?}

\item{include_crime}{Boolean, default \code{TRUE}. Include the Crime domain?}

\item{include_barriers}{Boolean, default \code{TRUE}. Include the Barriers to
Housing and Services domain?}

\item{include_environment}{Boolean, default \code{TRUE}. Include the
Living Environment Deprivation domain?}
}
\value{
A logical vector
}
\description{
All domains are included by default
}
