% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-wrenches.R
\name{g.class}
\alias{g.class}
\title{A function to diagnose differences between two allele vectors}
\usage{
g.class(x, y, flip_strand = TRUE)
}
\arguments{
\item{x}{a first allele vector}

\item{y}{a second allele vector}

\item{flip_strand}{If TRUE, it will assume that complementary and reverse complementary alleles are possible. This will also remove ambiguous alleles.}
}
\value{
a diagnostic vector
}
\description{
g.class, as found in annotSnpStats package (github.com/chr1swallace/annotSnpStats/)
}
