% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-wrenches.R
\name{g.complement}
\alias{g.complement}
\title{A function to find the complement to a vector of alleles}
\usage{
g.complement(x)
}
\arguments{
\item{x}{an allele vector}
}
\value{
a complement vector of alleles
}
\description{
Just like the one in annotSnpStats package (github.com/chr1swallace/annotSnpStats/)
}
