% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-wrenches.R
\name{g.rev}
\alias{g.rev}
\title{A function to flip alleles}
\usage{
g.rev(x, sep = "/")
}
\arguments{
\item{x}{an allele vector}

\item{sep}{the separator. Default: "/"}
}
\value{
a vector of flipped alleles
}
\description{
g.rev, as found in annotSnpStats package (github.com/chr1swallace/annotSnpStats/)
}
