% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imd_lookup.R
\name{get_imd_lookup}
\alias{get_imd_lookup}
\title{Import ranks and deciles for all English LSOAs for the 2025 IMD}
\usage{
get_imd_lookup()
}
\value{
A tibble with columns for LSOA codes and names, LAD codes and names,
and IMD ranks and deciles
}
\description{
Import ranks and deciles for all English LSOAs for the 2025 IMD
}
