% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transformed_scores.R
\name{get_overall_scores}
\alias{get_overall_scores}
\title{Download the overall IMD score for each LSOA}
\usage{
get_overall_scores()
}
\value{
A tibble with columns for LSOA codes and names, LAD codes and names,
and the overall score for each LSOA
}
\description{
Download the overall IMD score for each LSOA
}
