% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transformed_scores.R
\name{get_transformed_scores}
\alias{get_transformed_scores}
\title{Download the transformed domain scores for each LSOA}
\usage{
get_transformed_scores()
}
\value{
A tibble with columns for LSOA codes and names, LAD codes and names,
and the transformed scores for each of the seven domains
}
\description{
Download the transformed domain scores for each LSOA
}
