% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_imd_score.R
\name{imd_weights}
\alias{imd_weights}
\title{The per-domain weightings for the 2025 IMD}
\usage{
imd_weights()
}
\value{
A numeric vector
}
\description{
The per-domain weightings for the 2025 IMD
}
\keyword{internal}
