% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-wrenches.R
\name{linORscale}
\alias{linORscale}
\title{Transform Beta and SE from linear to logg-odds scale}
\usage{
linORscale(beta, se, N0, N1)
}
\arguments{
\item{beta}{a numeric vector of Beta in linear scale.}

\item{se}{a numeric vector of SE in linear scale.}

\item{N0}{number of controls in the study.}

\item{N1}{number of cases in the study.}
}
\value{
a list containing beta and se.beta in log-odds scale.
}
\description{
Convert effect sizes and standard error from linear to log-odds scale
}
\details{
Binary traits in case-control studies can be analysed by BOLT-LMM, which uses a linear model.
This function provides an approximation from linear to log-odds scale for Beta and SE, 
as described in BOLT-LMM manual (https://data.broadinstitute.org/alkesgroup/BOLT-LMM/).
}
