% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-wrenches.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot colour palette}
\usage{
plot_palette(palette)
}
\arguments{
\item{palette}{a palette vector with hexadecimal colour codes}
}
\value{
a plot showing the palette colours
}
\description{
Easily plot palette to test colour combinations
}
\details{
If you have a palette vector (with hexadecimal colour codes), you can simply pass it to this function to visualise it
}
