% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr-wrenches.R
\name{sdY.correction}
\alias{sdY.correction}
\title{Estimate trait variance, internal function}
\usage{
sdY.correction(beta, se, maf, n)
}
\arguments{
\item{beta}{vector of coefficients}

\item{se}{vector of standard error of coefficients}

\item{maf}{vector of MAF (same length as beta and se)}

\item{n}{sample size}
}
\value{
a list containing adjusted Beta and SE by estimated standard deviation of Y
}
\description{
Estimate trait standard deviation given vectors of variance of coefficients, MAF and sample size, and correct BETA and SE.
}
\details{
For studies of quantitative traits, measurements can come in different scales, leading to 
differences in standard deviation of the effects.
This function estimates the trait standard deviation from MAF and sample size, and by dividing beta and se 
by this estimate we can make var.beta = 1, thus making different datasets of quantitative traits in different
scales comparable.

Estimate is based on var(beta-hat) = var(Y) / (n * var(X))
var(X) = 2*maf*(1-maf)
so we can estimate var(Y) by regressing n*var(X) against 1/var(beta)

This is an adaptation of sdY.est function, by Chris Wallace
}
\author{
Guillermo Reales, Chris Wallace
}
