% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_imd_score.R
\name{calculate_imd_score}
\alias{calculate_imd_score}
\title{Calculate IMD score from domain scores}
\usage{
calculate_imd_score(scores_tbl, domains = choose_domains())
}
\arguments{
\item{scores_tbl}{A table containing transformed domain scores, as produced
by \link{get_transformed_scores}}

\item{domains}{A Boolean vector, best produced by \link{choose_domains}}
}
\description{
Calculate IMD score from domain scores
}
