% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_imd_score.R
\name{choose_domains}
\alias{choose_domains}
\title{Helper function to select domains to include in calculation of IMD overall}
\usage{
choose_domains(
  include_income = TRUE,
  include_employment = TRUE,
  include_education = TRUE,
  include_health = TRUE,
  include_crime = TRUE,
  include_barriers = TRUE,
  include_living_environment = TRUE
)
}
\arguments{
\item{include_income}{Boolean, default \code{TRUE}. Include the Income domain?}

\item{include_employment}{Boolean, default \code{TRUE}. Include the Employment
domain?}

\item{include_education}{Boolean, default \code{TRUE}. Include the Education
domain?}

\item{include_health}{Boolean, default \code{TRUE}. Include the Health domain?}

\item{include_crime}{Boolean, default \code{TRUE}. Include the Crime domain?}

\item{include_barriers}{Boolean, default \code{TRUE}. Include the Barriers domain?}

\item{include_living_environment}{Boolean, default \code{TRUE}. Include the
Living Environment domain?}
}
\value{
A Boolean vector
}
\description{
All domains are included by default
}
