% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_imd_score.R
\name{imd_weights}
\alias{imd_weights}
\title{The per-domain weightings for the 2025 IMD}
\usage{
imd_weights()
}
\value{
A numeric vector of length 7
}
\description{
In order: income, employment, education, health deprivation and disability,
crime, barriers to housing and services, living environment deprivation
}
